Styling Changes
This commit adds a footer to the bottom of the page. Users will now see a programming/hacking animation at the bottom of their window. This addition will hopefully make the app feel more fun. Eventually, I might even add a timer in the upper right hand corner of the screen so users can track how long it takes them tosolve a challenge.